Added some pages and modules. Restructured directories.
