+++ src/App.js
@@ -1,7 +1,7 @@
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import './App.css';
-import Footer from './modules/footer';
-import NavBar from './modules/navBar';
+import Footer from './modules/footer/';
+import NavBar from './modules/navBar/';
 import Dashboard from './pages/dashboard';
 import HomePage from './pages/home';
