+++ src/modules/banner/component.js
@@ -1,0 +1,10 @@
+const Banner = ({ text, children }) => {
+  return ( 
+    <div className="banner">
+      <h1>{ text }</h1>
+      { children }
+    </div>
+   );
+}
+ 
+export default Banner;
