+++ src/modules/banner/index.js
@@ -1,0 +1,3 @@
+import Banner from "./component";
+
+export default Banner;
