+++ src/modules/footer/component.js
@@ -1,4 +1,4 @@
-import { APP_NAME } from '../constants';
+import { APP_NAME } from '../../constants';
 
 const Footer = () => {
   return (
