+++ src/modules/footer/index.js
@@ -1,0 +1,3 @@
+import Footer from "./component";
+
+export default Footer;
