+++ src/modules/heading/component.js
@@ -1,0 +1,7 @@
+const Heading = ({ title }) => {
+  return ( 
+    <h1>{title}</h1>
+   );
+}
+ 
+export default Heading;
