+++ src/modules/heading/index.js
@@ -1,0 +1,3 @@
+import Heading from "./component";
+
+export default Heading;
