+++ src/modules/navBar/component.js
@@ -1,5 +1,5 @@
 import { Link } from 'react-router-dom';
-import { APP_NAME } from '../constants';
+import { APP_NAME } from '../../constants';
 
 const NavBar = () => {
   return (
