+++ src/modules/navBar/index.js
@@ -1,0 +1,3 @@
+import NavBar from "./component";
+
+export default NavBar;
