+++ src/pages/home/index.js
@@ -1,6 +1,13 @@
+import Banner from "../../modules/banner/"
+import Heading from "../../modules/heading/"
+
 const HomePage = () => {
   return (
-    <div>Hello World</div>
+    <div>
+      <Banner text="Heading 1">
+        <Heading title="Heading 2" />
+      </Banner>
+    </div>
   );
 }
